avoid using readonly on a field updated with onchange
