+++ sale_order_secondary_unit_price/i18n/ja.po
@@ -1,0 +1,27 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* sale_order_secondary_unit_price
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2019-09-11 14:58+0000\n"
+"PO-Revision-Date: 2019-09-11 14:58+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: sale_order_secondary_unit_price
+#: model:ir.model,name:sale_order_secondary_unit_price.model_sale_order_line
+msgid "Sales Order Line"
+msgstr "販売オーダ明細"
+
+#. module: sale_order_secondary_unit_price
+#: model:ir.model.fields,field_description:sale_order_secondary_unit_price.field_sale_order_line__secondary_unit_price
+msgid "Secondary Unit Price"
+msgstr "単価(副単位)"
+
