+++ sale_order_secondary_unit_price/models/__init__.py
@@ -1,1 +1,1 @@
-from . import sale_order
+from . import sale_order_line
