+++ sale_order_secondary_unit_price/views/sale_order_views.xml
@@ -7,13 +7,10 @@
         <field name="groups_id" eval="[(4, ref('uom.group_uom'))]"/>
         <field name="arch" type="xml">
             <xpath expr="//field[@name='order_line']/form//field[@name='price_subtotal']" position="after">
-                <field name="secondary_unit_price" attrs="{'readonly': [('parent.state', 'in', ('done', 'cancel'))]}"/>
+                <field name="secondary_unit_price" attrs="{'readonly':[('parent.state','in',('done','cancel'))]}"/>
             </xpath>
             <xpath expr="//field[@name='order_line']/tree//field[@name='product_uom_qty']" position="before">
-                <field name="secondary_unit_price" attrs="{'readonly': [('parent.state', 'in', ('done', 'cancel'))]}"/>
-            </xpath>
-            <xpath expr="//field[@name='order_line']/tree//field[@name='price_unit']" position="attributes">
-                <attribute name="attrs">{'readonly': [('secondary_unit_price', '!=', 0.0)]}</attribute>
+                <field name="secondary_unit_price" attrs="{'readonly':[('parent.state','in',('done','cancel'))]}"/>
             </xpath>
         </field>
     </record>
